refactor data grid outside of vaccine info. Rename component.
