+++ frontend/Components/VaccineInfo/AccordionItem.jsx
