+++ frontend/Components/VaccineInfo/Card.jsx
